// TeaterAdmin/src/app/model/hall.ts
/**
 * TicketService API
 * Ticketordering by date
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { Theater } from './theater';

   export interface Hall {
    hallNum?: number;
    seats?: number;
    rows?: number;
    theater?: Theater;
  }